Chore: move internal rules out of lib/
This moves project-specific rules out of the `lib/` directory and
into `tools`. This prevents the rules from being included in the
published npm package, and it also allows the rule files to use
devDependencies without linter errors (since the devDependencies are
    always present when running the rules).
