+++ tools/internal-rules/internal-consistent-docs-description.js
