+++ tools/internal-rules/internal-no-invalid-meta.js
